feat: show period total time in summary card, #1752
